chore: remove GitPod support (#705)
As noted in https://github.com/renovatebot/renovate/discussions/37856,
we (the maintainers) have not used GitPod for maintaining the docs.

Additionally, as the company Ona (previously known as GitPod) is
pivoting, we should remove GitPod support, and we removed support in the
Renovate project in https://github.com/renovatebot/renovate/pull/22274,
we should remove GitPod support in our docs, too.

As per #704, we don't (yet) have support for GitHub Codespaces, but
should add it at some point.