feat: track Renovate dependency as a Git tag (#739)
By tracking the dependency as a Git tag, we can make sure that we're
always referencing a known release, and make it more straightforward to
move towards a model noted in #108 in the future.

This will also make it so we can wire in the Renovate version to build
our docs, highlighting the version of Renovate the docs are for.

This also intentionally downgrades the tag version of the docs, to make
sure that Renovate attempts to upgrade it.