feat: display Renovate tag in page footer (#741)
As part of wiring in the version of the CLI into the docs build process,
we should also show this version to users, so they're aware what version
of the docs they're looking at.

To do this, we can wire in the newly added `RENOVATE_VERSION`
environment variable that MkDocs will use.

We should also make sure that we're picking up the version in our
`Makefile` to keep it updated.